requirements.txt
alembic==1.8.1
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
zipp==3.11.0
